Added a basic, thread safe profiler for tracking time spent in any function. Not sure what the units are. Sure as hell not seconds. Left in a few macros to do the conversion if you are so inclined.
Also profiled CometPreprocess functions. Summary of the results provided at program termination. 

To use the profiler elsewhere in the code, simply start a timer (Profiler::StartTimer), give it a name to store the results, and collect the return process ID. When ready, stop the timer using the process ID. If you start multiple timers with the same name, results will be summed. Use different names to keep profiles separate. See CometPreprocess for examples where timers are started at the entrance and exit of most functions.

To get around idle time while in a nested function, stop a timer, process the nested function, then start the timer again when it returns.

git-svn-id: https://svn.code.sf.net/p/comet-ms/code@375 299d2d02-8f3b-485e-a07e-71e8e5e16577